requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
streamlit>=1.35.0
opencv-python>=4.8.0
Pillow>=10.0.0
neurokit2>=0.2.7
wfdb>=4.1.0
tqdm>=4.65.0
pdf2image>=1.16.0
pypdf>=3.0.0
plotly>=5.18.0
